backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.2
numpy==1.26.4
opencv-python==4.9.0.80
qrcode==7.4.2
Pillow==10.3.0
face-recognition==1.3.0
